Support for tracing back the edges of the MST. The fnction signature recall in some ways the one  supplied by the boost library. Instead of the recursive approach proposed in the paper here I decide to exploit 2 vectors: one keeps track of the edgeID from the start of the program (onGPU.edgeID), while the other save the result of the MST (onGPU.edgeIDresult). To permit saving procedure an additional integer has been added to DatastructuresOnGpu (savedEdges). The verifyMST procedure has been changed accordingly in order to check the additional constraint of the cost.

+++ FastMST/NVEcell.cu
@@ -6,8 +6,6 @@ __host__ __device__ bool operator < (const NVEcell &lhs, const NVEcell &rhs)
 	return (lhs.cell < rhs.cell);
 };
 
-
-
 __host__ __device__ void NVEcell::setSource(unsigned int s) {
 	long long tmp = s;
 	tmp <<= WEIGHT_SIZE + VERTEX_SIZE;
@@ -40,12 +38,8 @@ __host__ __device__ unsigned int NVEcell::getDestination() {
 }
 
 __host__ __device__ unsigned int NVEcell::getWeight() {
-	unsigned int tmp = cell;
+	unsigned int tmp = (unsigned int) cell;
 	int mask = createMask(0, WEIGHT_SIZE);
 
 	return tmp & mask;
 }
-
-
-
-
